Generify timezone, getter function use, change twitter to x
This commit changes timezone handling so that it works based off of UTC
instead of system time, this means that the app will work even when the
host machine timezone changes.
Changed to getter functions for the main times used throughout for
better maintainability.
All references to Twitter have changed to X.